Change link, slot, crate to uint32

+++ include/fdreadoutlibs/wib2/WIB2FrameProcessor.hpp
@@ -753,9 +753,9 @@ private:
 
   std::atomic<bool> m_add_hits_tphandler_thread_should_run;
 
-  uint8_t m_link; // NOLINT(build/unsigned)
-  uint8_t m_slot_no;  // NOLINT(build/unsigned)
-  uint8_t m_crate_no; // NOLINT(build/unsigned)
+  uint32_t m_link; // NOLINT(build/unsigned)
+  uint32_t m_slot_no;  // NOLINT(build/unsigned)
+  uint32_t m_crate_no; // NOLINT(build/unsigned)
 
   std::shared_ptr<detchannelmaps::TPCChannelMap> m_channel_map;
